+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
 ## [unreleased]
+## Global Accelerator
+- added a new module to launch an AWS Global Accelerator, [#104](https://github.com/dbl-works/terraform/pull/104)
 
 ## ElastiCache
 - added `data_tiering_enabled` to the available options, defaulting to `false`